+++ src/controllers/Serializer.ts
@@ -46,7 +46,7 @@ import { getListFromObject, getMethodOrThrow } from './utils';
  *
  *     response.writeHead(200, {
  *       'Content-Length': String(jsonResponse.length),
- *       'Content-Type': 'application/json; charset=utf-8'
+ *       'Content-Type': 'application/json; charset=UTF-8'
  *     });
  *     response.write(jsonResponse)
  *   }